# backend/src/main/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/rentals/{id}":{"get":{"tags":["rental-controller"],"operationId":"getRental","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RentalDto"}}}}}},"put":{"tags":["rental-controller"],"operationId":"updateRental","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RentalDto"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["rental-controller"],"operationId":"deleteRental","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/customers/{id}":{"get":{"tags":["customer-controller"],"operationId":"getCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerDto"}}}}}},"put":{"tags":["customer-controller"],"operationId":"updateCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDto"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["customer-controller"],"operationId":"deleteCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/cars/{id}":{"get":{"tags":["car-controller"],"operationId":"getCar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CarDto"}}}}}},"put":{"tags":["car-controller"],"operationId":"updateCar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarDto"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["car-controller"],"operationId":"deleteCar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/rentals/create":{"post":{"tags":["rental-controller"],"operationId":"createRental","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RentalDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/customers/create":{"post":{"tags":["customer-controller"],"operationId":"createCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/cars/create":{"post":{"tags":["car-controller"],"operationId":"createCar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/rentals/":{"get":{"tags":["rental-controller"],"operationId":"getRentals","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RentalDto"}}}}}}}},"/api/customers/":{"get":{"tags":["customer-controller"],"operationId":"getCustomers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CustomerDto"}}}}}}}},"/api/cars/":{"get":{"tags":["car-controller"],"operationId":"getCars","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CarDto"}}}}}}}}},"components":{"schemas":{"RentalDto":{"type":"object","properties":{"rentalId":{"type":"integer","format":"int64"},"carId":{"type":"integer","format":"int64"},"carLicensePlate":{"type":"string"},"carBrandAndModel":{"type":"string"},"customerId":{"type":"integer","format":"int64"},"customerFullName":{"type":"string"},"customerEmail":{"type":"string"},"rentalBegin":{"type":"string","format":"date-time"},"rentalEnd":{"type":"string","format":"date-time"},"returnDate":{"type":"string","format":"date-time"},"mileageBegin":{"type":"integer","format":"int32"},"mileageEnd":{"type":"integer","format":"int32"},"description":{"type":"string"},"createdWhen":{"type":"string","format":"date-time"}}},"CustomerDto":{"type":"object","properties":{"customerId":{"type":"integer","format":"int64"},"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"birthDate":{"type":"string","format":"date"},"drivingLicense":{"type":"string"},"streetName":{"type":"string"},"houseNumber":{"type":"string"},"postCode":{"type":"string"},"city":{"type":"string"},"country":{"type":"string"},"createdWhen":{"type":"string","format":"date-time"}}},"CarDto":{"type":"object","properties":{"carId":{"type":"integer","format":"int64"},"licensePlate":{"type":"string"},"brand":{"type":"string"},"model":{"type":"string"},"constructionYear":{"type":"string"},"color":{"type":"string"},"chassisNumber":{"type":"string"},"mileage":{"type":"integer","format":"int32"},"status":{"type":"string"},"category":{"type":"string"},"dailyPrice":{"type":"number"},"image64":{"type":"string"}}}}}}